Enable boot without watchdog
This patch allows booting an image without arming watchdog timer
when "watchdog=0" parameter is passed to the to the module.

This covers the use case when availability of a hardware driver is
unknown ahead of time and can only be established upon first boot.
The code that runs on the first boot can then identify if supported
hardware watchdog is available, the required driver is present in
the kernel and either update boot configuration re-arming the timer
or issuing a warning about unsupported hardware. This would also
allow for booting images in virtual environments.